Preserve schema of input dataframe
Plugin does not fully export formatted text if the formatted text is made up of numbers.  Especially problematic if the string contains a number starting with '0'.   So, '019' exports as 19 which is not correct.

Change the get_dataframe call to so that infer_with_pandas = False

This more closely matches the native "Export xlsx" functionality in DSS.